+++ web/js/ckeditor/plugins/autocolorize/lang/de.js
@@ -1,0 +1,3 @@
+CKEDITOR.plugins.setLang( 'autocolorize', 'de', {
+    buttonTitle: 'Eingegebenen Text farblich hervorherben'
+} );