+++ web/js/ckeditor/plugins/autocolorize/lang/en.js
@@ -1,0 +1,3 @@
+CKEDITOR.plugins.setLang( 'autocolorize', 'en', {
+    buttonTitle: 'Colorize entered text'
+} );